yeah no idk if wpi can support this
it cant be the code. it has to be just the roborio not being able to handle this or something. literally added one unrelated thing, went from 7 moves to 5 moves without changing the amount of moves. then unchanged the unrelated thing and it didn't fix the problem. its been super unreliable and random, my code works. legit has to