2102 gavin is boss
literally solved what i couldnt for so long

it was the damn motors not resetting. i knew it. thought it was me cuz i aint the best, but it wasnt the code lmao. kind of.

switching sin and cos made it go up and down instead of left and right, hopefully that is its tru config